nothing

+++ src/simulation/Mass.java
@@ -40,21 +40,23 @@ public class Mass extends PhysicalObjectRect{
 //		y = springHookY+myHeight/2;
 //	}
 	
-	int prevmousex = 0, prevmousey = 0;
+	double prevmousex = 0, prevmousey = 0;
 	
 	public void move()	{
+		
 		if(myEngine.getMouseButton(1) && (myEngine.getMouseX() >= x-massWidth*2 && 
 				myEngine.getMouseX() <= x+massWidth*2) && 
 				(myEngine.getMouseY() >= y-massHeight*2 &&
 				myEngine.getMouseY() <= y+massHeight*2))	{
 			x = myEngine.getMouseX()-massWidth/2;
 			y = myEngine.getMouseY()-massHeight/2;
-			xspeed = (x-prevmousex)/10;
-			yspeed = (y-prevmousey)/10;
+			xspeed = (x-prevmousex)/10.0;
+			yspeed = (y-prevmousey)/10.0;
 			prevmousex = myEngine.getMouseX();
 			prevmousey = myEngine.getMouseY();
 		}
 	}
+	
 	
 	
 }
